+++ internal/server/main.go
@@ -6,6 +6,7 @@
 package server
 
 import (
+	"context"
 	"net/http"
 
 	"github.com/circonus-labs/circonus-agent/internal/config"
@@ -20,6 +21,7 @@ import (
 
 // Server defines the listening servers
 type Server struct {
+	ctx       context.Context
 	logger    zerolog.Logger
 	plugins   *plugins.Plugins
 	svrHTTP   *http.Server
@@ -28,8 +30,9 @@ type Server struct {
 }
 
 // New creates a new instance of the listening servers
-func New(p *plugins.Plugins, ss *statsd.Server) *Server {
+func New(ctx context.Context, p *plugins.Plugins, ss *statsd.Server) *Server {
 	s := Server{
+		ctx:       ctx,
 		logger:    log.With().Str("pkg", "server").Logger(),
 		plugins:   p,
 		statsdSvr: ss,
@@ -57,7 +60,7 @@ func (s *Server) Start() error {
 	}
 	// Manual waitgroup for the situation where both servers are started;
 	// one fails and the other doesn't - wg.Wait() would block.
-	// The desired behavior is for an error in *either* to abort the process.
+	// The desired behavior is for an error in *either* to abort the process (somewhat cleanly).
 	// there is probably a more idiomatic way to handle this...
 	numDone := 0
 	expected := 0
@@ -72,7 +75,7 @@ func (s *Server) Start() error {
 			defer s.svrHTTP.Close()
 			s.logger.Info().Str("listen", s.svrHTTP.Addr).Msg("Starting")
 			if err := s.svrHTTP.ListenAndServe(); err != nil {
-				if err.Error() != "http: Server closed" {
+				if err != http.ErrServerClosed {
 					ec <- err
 				}
 			}
@@ -90,24 +93,44 @@ func (s *Server) Start() error {
 			keyFile := viper.GetString(config.KeySSLKeyFile)
 			s.logger.Info().Str("listen", s.svrHTTPS.Addr).Msg("SSL starting")
 			if err := s.svrHTTPS.ListenAndServeTLS(certFile, keyFile); err != nil {
-				ec <- errors.Wrap(err, "SSL server")
+				if err != http.ErrServerClosed {
+					ec <- errors.Wrap(err, "SSL server")
+				}
 			}
 			done <- 1
 		}()
 	}
 
-	if expected > 0 {
+	for numDone < expected {
 		select {
+		case <-s.ctx.Done():
+			return nil
 		case err := <-ec:
 			return err
 		case <-done:
 			numDone++
 			s.logger.Debug().Int("done", numDone).Msg("completed")
-			if numDone == expected {
-				break
-			}
 		}
 	}
 
 	return nil
+}
+
+// Stop the servers
+func (s *Server) Stop() {
+	if s.svrHTTP != nil {
+		s.logger.Debug().Msg("Stopping HTTP server")
+		err := s.svrHTTP.Close()
+		if err != nil {
+			s.logger.Warn().Err(err).Msg("Closing HTTP server")
+		}
+	}
+
+	if s.svrHTTPS != nil {
+		s.logger.Debug().Msg("Stopping HTTPS server")
+		err := s.svrHTTPS.Close()
+		if err != nil {
+			s.logger.Warn().Err(err).Msg("Closing HTTPS server")
+		}
+	}
 }