+++ internal/server/main_test.go
@@ -6,6 +6,7 @@
 package server
 
 import (
+	"context"
 	"errors"
 	"testing"
 	"time"
@@ -22,7 +23,7 @@ func TestServerHTTP(t *testing.T) {
 
 	t.Log("No config")
 	{
-		s := New(nil, nil)
+		s := New(context.Background(), nil, nil)
 		if s.svrHTTP != nil {
 			t.Fatal("expected nil")
 		}
@@ -31,7 +32,7 @@ func TestServerHTTP(t *testing.T) {
 	t.Log("With config")
 	{
 		viper.Set(config.KeyListen, ":2609")
-		s := New(nil, nil)
+		s := New(context.Background(), nil, nil)
 		if s.svrHTTP == nil {
 			t.Fatal("expected NOT nil")
 		}
@@ -45,7 +46,7 @@ func TestServerHTTPS(t *testing.T) {
 
 	t.Log("No config")
 	{
-		s := New(nil, nil)
+		s := New(context.Background(), nil, nil)
 		if s.svrHTTPS != nil {
 			t.Fatal("expected nil")
 		}
@@ -54,7 +55,7 @@ func TestServerHTTPS(t *testing.T) {
 	t.Log("With config")
 	{
 		viper.Set(config.KeySSLListen, ":2610")
-		s := New(nil, nil)
+		s := New(context.Background(), nil, nil)
 		viper.Reset()
 		if s.svrHTTPS == nil {
 			t.Fatal("expected NOT nil")
@@ -70,9 +71,9 @@ func TestRunServers(t *testing.T) {
 	t.Log("HTTP")
 	{
 		viper.Set(config.KeyListen, ":65111")
-		s := New(nil, nil)
+		s := New(context.Background(), nil, nil)
 		time.AfterFunc(2*time.Second, func() {
-			s.svrHTTP.Close()
+			s.Stop()
 		})
 		if err := s.Start(); err != nil {
 			t.Fatalf("expected NO error, got (%v)", err)
@@ -83,7 +84,7 @@ func TestRunServers(t *testing.T) {
 	t.Log("HTTPS (no cert/key config)")
 	{
 		viper.Set(config.KeySSLListen, ":65222")
-		s := New(nil, nil)
+		s := New(context.Background(), nil, nil)
 		expectedErr := errors.New("SSL server: open : no such file or directory")
 		err := s.Start()
 		if err == nil {
@@ -93,7 +94,7 @@ func TestRunServers(t *testing.T) {
 		if err.Error() != expectedErr.Error() {
 			t.Fatalf("expected (%s) got (%v)", expectedErr, err)
 		}
-		s.svrHTTPS.Close()
+		s.Stop()
 		viper.Reset()
 	}
 
@@ -101,7 +102,7 @@ func TestRunServers(t *testing.T) {
 	{
 		viper.Set(config.KeySSLListen, ":65223")
 		viper.Set(config.KeySSLCertFile, "testdata/missing.crt")
-		s := New(nil, nil)
+		s := New(context.Background(), nil, nil)
 		expectedErr := errors.New("SSL server: open testdata/missing.crt: no such file or directory")
 		err := s.Start()
 		if err == nil {
@@ -111,7 +112,7 @@ func TestRunServers(t *testing.T) {
 		if err.Error() != expectedErr.Error() {
 			t.Fatalf("expected (%s) got (%v)", expectedErr, err)
 		}
-		s.svrHTTPS.Close()
+		s.Stop()
 		viper.Reset()
 	}
 
@@ -120,7 +121,7 @@ func TestRunServers(t *testing.T) {
 		viper.Set(config.KeySSLListen, ":65224")
 		viper.Set(config.KeySSLCertFile, "testdata/cert.crt")
 		viper.Set(config.KeySSLKeyFile, "testdata/missing.key")
-		s := New(nil, nil)
+		s := New(context.Background(), nil, nil)
 		expectedErr := errors.New("SSL server: open testdata/missing.key: no such file or directory")
 		err := s.Start()
 		if err == nil {
@@ -130,7 +131,7 @@ func TestRunServers(t *testing.T) {
 		if err.Error() != expectedErr.Error() {
 			t.Fatalf("expected (%s) got (%v)", expectedErr, err)
 		}
-		s.svrHTTPS.Close()
+		s.Stop()
 		viper.Reset()
 	}
 
@@ -139,7 +140,7 @@ func TestRunServers(t *testing.T) {
 		viper.Set(config.KeySSLListen, ":65225")
 		viper.Set(config.KeySSLCertFile, "testdata/cert.crt")
 		viper.Set(config.KeySSLKeyFile, "testdata/key.key")
-		s := New(nil, nil)
+		s := New(context.Background(), nil, nil)
 		expectedErr := errors.New("SSL server: tls: failed to find any PEM data in certificate input")
 		err := s.Start()
 		if err == nil {
@@ -149,7 +150,7 @@ func TestRunServers(t *testing.T) {
 		if err.Error() != expectedErr.Error() {
 			t.Fatalf("expected (%s) got (%v)", expectedErr, err)
 		}
-		s.svrHTTPS.Close()
+		s.Stop()
 		viper.Reset()
 	}
 }