+++ internal/statsd/main_test.go
@@ -6,6 +6,7 @@
 package statsd
 
 import (
+	"context"
 	"errors"
 	"testing"
 
@@ -22,18 +23,25 @@ func TestFlush(t *testing.T) {
 	t.Log("Disabled")
 	{
 		viper.Set(config.KeyStatsdDisabled, true)
-		s := New()
+		s, err := New(context.Background())
+		if err != nil {
+			t.Fatalf("expected NO error, got (%s)", err)
+		}
 		viper.Reset()
 
-		if s != nil {
-			t.Fatalf("expected nil")
+		if err := s.Start(); err != nil {
+			t.Fatalf("expected nil, got (%s)", err)
 		}
+		s.Stop()
 	}
 
 	t.Log("Flush (disabled)")
 	{
 		viper.Set(config.KeyStatsdDisabled, true)
-		s := New()
+		s, err := New(context.Background())
+		if err != nil {
+			t.Fatalf("expected NO error, got (%s)", err)
+		}
 		metrics := s.Flush()
 		viper.Reset()
 
@@ -44,7 +52,10 @@ func TestFlush(t *testing.T) {
 
 	t.Log("Flush (no stats)")
 	{
-		s := New()
+		s, err := New(context.Background())
+		if err != nil {
+			t.Fatalf("expected NO error, got (%s)", err)
+		}
 		metrics := s.Flush()
 
 		if metrics == nil {
@@ -73,7 +84,10 @@ func TestGetMetricDest(t *testing.T) {
 			{"group.foo", "host", "group.foo"},
 			{"foo", "host", "foo"},
 		}
-		s := New()
+		s, err := New(context.Background())
+		if err != nil {
+			t.Fatalf("expected NO error, got (%s)", err)
+		}
 		for _, d := range dtest {
 			t.Logf("%s -> %s", d.metricName, d.expectedDest)
 			dest, metric := s.getMetricDestination(d.metricName)
@@ -101,7 +115,10 @@ func TestGetMetricDest(t *testing.T) {
 		}
 		viper.Set(config.KeyStatsdHostPrefix, "host.")
 		viper.Set(config.KeyStatsdGroupPrefix, "group.")
-		s := New()
+		s, err := New(context.Background())
+		if err != nil {
+			t.Fatalf("expected NO error, got (%s)", err)
+		}
 		viper.Reset()
 		for _, d := range dtest {
 			t.Logf("%s -> %s", d.metricName, d.expectedDest)
@@ -129,7 +146,10 @@ func TestGetMetricDest(t *testing.T) {
 			{"foo", "host", "foo"},
 		}
 		viper.Set(config.KeyStatsdGroupPrefix, "group.")
-		s := New()
+		s, err := New(context.Background())
+		if err != nil {
+			t.Fatalf("expected NO error, got (%s)", err)
+		}
 		viper.Reset()
 		for _, d := range dtest {
 			t.Logf("%s -> %s", d.metricName, d.expectedDest)
@@ -157,7 +177,10 @@ func TestGetMetricDest(t *testing.T) {
 			{"foo", "group", "foo"},
 		}
 		viper.Set(config.KeyStatsdHostPrefix, "host.")
-		s := New()
+		s, err := New(context.Background())
+		if err != nil {
+			t.Fatalf("expected NO error, got (%s)", err)
+		}
 		viper.Reset()
 		for _, d := range dtest {
 			t.Logf("%s -> %s", d.metricName, d.expectedDest)
@@ -177,7 +200,10 @@ func TestParseMetric(t *testing.T) {
 
 	zerolog.SetGlobalLevel(zerolog.Disabled)
 
-	s := New()
+	s, err := New(context.Background())
+	if err != nil {
+		t.Fatalf("expected NO error, got (%s)", err)
+	}
 
 	t.Log("Blank")
 	{