+++ CHANGELOG.md
@@ -1,4 +1,9 @@
+# v0.0.3
+    * integrate context
+    * cleaner shutdown handling
 # v0.0.2
     * move `circonus-agentd` binary to `sbin/circonus-agentd`
+    * refactor (plugins, server, reverse, statsd)
+    * add agent package
 # v0.0.1
     * Initial development working release